Allows for queuing of prompts.
Instead of holding up the prompt editor from being used until the
last dispatched prompt request is done, this change will queue the
prompt in the history list for processing (not sure if history list
is a proper name anymore). This allows for the next experiment to
be off-loaded one's brain and entered in the prompt editor as soon
as possible.